+++ CHANGELOG.md
@@ -4,9 +4,12 @@
 ## 0.1.0.1 - unreleased
 
 -   Add to playlist prompt now updates for new playlists
--   Playlists now sorted alphabetically
+-   Playlists now sorted alphabetically -- kill orphan
 -   List of playlists keeps cursor position after edit
--   `\`` toggles album sort order in library
+-   Library now shows track number on songs
+-   Library now uses album artist on left column
+-   Library now displays and sorts by album year by default
+-   `\`` toggles album sort order in library (between year and alphabetical)
 
 ## 0.1.0.0 - 2020-12-12