+++ CHANGELOG.md
@@ -1,16 +1,19 @@
 
 # Revision history for hum
 
-## 0.1.0.1 - unreleased
+## 0.2.0.0 - unreleased
 
 -   Add to playlist prompt now updates for new playlists
--   Playlists now sorted alphabetically -- kill orphan
+-   Playlists now sorted alphabetically
 -   List of playlists keeps cursor position after edit
 -   Library now shows track number on songs
 -   Library now uses album artist on left column
 -   Library now displays and sorts by album year by default
 -   `\`` toggles album sort order in library (between year and alphabetical)
 -   Help can now be scrolled and searched with case insensitive POSIX regex
+-   Fix add-to-playlist prompt not quitting when adding to old playlist
+-   Playlists can now be renamed
+
 
 ## 0.1.0.0 - 2020-12-12